switched to using static_cast from C-style casts; fixed renderer being set to nullptr instead of compare

+++ include/engine/button.h
@@ -126,7 +126,7 @@ public:
 
     void operator=(const ButtonGroup& copy);
 
-    Button &operator[](int iterindex);
+    Button &operator[](long long unsigned int iterindex);
 
     Button &operator[](string uid);
